+++ src/sign/signature.rs
@@ -706,9 +706,9 @@ impl<C: CipherSuite> ThresholdSignature<C> {
 mod test {
     use super::*;
 
-    use crate::dkg::{Coefficients, EncryptedSecretShare, Participant};
+    use crate::dkg::test::do_keygen;
+    use crate::dkg::Participant;
     use crate::dkg::{DistributedKeyGeneration, RoundOne};
-    use crate::keys::DiffieHellmanPrivateKey;
     use crate::sign::{generate_commitment_share_lists, PublicCommitmentShareList};
     use crate::testing::Secp256k1Sha256;
     use crate::{FromBytes, ToBytes};
@@ -717,187 +717,6 @@ mod test {
 
     use ark_ff::UniformRand;
     use rand::rngs::OsRng;
-
-    fn do_keygen(
-        n1: u32,
-        t1: u32,
-        n2: Option<u32>,
-        t2: Option<u32>,
-    ) -> FrostResult<
-        Secp256k1Sha256,
-        (
-            ThresholdParameters<Secp256k1Sha256>,
-            Vec<IndividualSigningKey<Secp256k1Sha256>>,
-            GroupVerifyingKey<Secp256k1Sha256>,
-            Option<ThresholdParameters<Secp256k1Sha256>>,
-            Option<Vec<IndividualSigningKey<Secp256k1Sha256>>>,
-        ),
-    > {
-        type Dkg<T> = DistributedKeyGeneration<T, Secp256k1Sha256>;
-
-        let params = ThresholdParameters::new(n1, t1);
-        let rng = OsRng;
-
-        let mut participants = Vec::<Participant<Secp256k1Sha256>>::new();
-        let mut coefficients = Vec::<Coefficients<Secp256k1Sha256>>::new();
-        let mut dh_secret_keys = Vec::<DiffieHellmanPrivateKey<Secp256k1Sha256>>::new();
-
-        for i in 1..=n1 {
-            let (p, c, dh_sk) = Participant::<Secp256k1Sha256>::new_dealer(params, i, rng)?;
-            participants.push(p);
-            coefficients.push(c);
-            dh_secret_keys.push(dh_sk);
-        }
-
-        let mut participants_encrypted_secret_shares: Vec<
-            BTreeMap<u32, EncryptedSecretShare<Secp256k1Sha256>>,
-        > = (0..n1).map(|_| BTreeMap::new()).collect();
-
-        let mut participants_states_1 = Vec::<Dkg<_>>::new();
-        let mut participants_states_2 = Vec::<Dkg<_>>::new();
-
-        for i in 0..n1 {
-            let (pi_state, _participant_lists) = Dkg::<_>::bootstrap(
-                params,
-                &dh_secret_keys[i as usize],
-                participants[i as usize].index,
-                &coefficients[i as usize],
-                &participants,
-                rng,
-            )?;
-            let pi_their_encrypted_secret_shares = pi_state.their_encrypted_secret_shares()?;
-            participants_encrypted_secret_shares[i as usize] =
-                pi_their_encrypted_secret_shares.clone();
-            participants_states_1.push(pi_state);
-        }
-
-        let mut p1_my_encrypted_secret_shares = Vec::<EncryptedSecretShare<Secp256k1Sha256>>::new();
-        for j in 0..n1 {
-            p1_my_encrypted_secret_shares.push(
-                participants_encrypted_secret_shares[j as usize]
-                    .get(&1)
-                    .unwrap()
-                    .clone(),
-            );
-        }
-        participants_states_2.push(
-            participants_states_1[0]
-                .clone()
-                .to_round_two(&p1_my_encrypted_secret_shares, rng)?
-                .0,
-        );
-
-        for i in 2..=n1 {
-            let mut pi_my_encrypted_secret_shares =
-                Vec::<EncryptedSecretShare<Secp256k1Sha256>>::new();
-            for j in 0..n1 {
-                pi_my_encrypted_secret_shares.push(
-                    participants_encrypted_secret_shares[j as usize]
-                        .get(&i)
-                        .unwrap()
-                        .clone(),
-                );
-            }
-
-            participants_states_2.push(
-                participants_states_1[(i - 1) as usize]
-                    .clone()
-                    .to_round_two(&pi_my_encrypted_secret_shares, rng)?
-                    .0,
-            );
-        }
-
-        let mut participants_secret_keys = Vec::<IndividualSigningKey<Secp256k1Sha256>>::new();
-        let (group_key, p1_sk) = participants_states_2[0].clone().finish()?;
-        participants_secret_keys.push(p1_sk);
-
-        for i in 2..=n1 {
-            let (_, pi_sk) = participants_states_2[(i - 1) as usize].clone().finish()?;
-
-            participants_secret_keys.push(pi_sk);
-        }
-
-        if let (Some(n2), Some(t2)) = (n2, t2) {
-            let new_params = ThresholdParameters::new(n2, t2);
-
-            let mut signers = Vec::<Participant<Secp256k1Sha256>>::new();
-            let mut signers_dh_secret_keys = Vec::<DiffieHellmanPrivateKey<Secp256k1Sha256>>::new();
-
-            for i in 1..=n2 {
-                let (p, dh_sk) = Participant::<Secp256k1Sha256>::new_signer(params, i, rng)?;
-
-                signers.push(p);
-                signers_dh_secret_keys.push(dh_sk);
-            }
-
-            let mut signers_encrypted_secret_shares: Vec<
-                Vec<EncryptedSecretShare<Secp256k1Sha256>>,
-            > = (0..n2).map(|_| Vec::new()).collect();
-
-            let mut signers_states_1 = Vec::<Dkg<_>>::new();
-            let mut signers_states_2 = Vec::<Dkg<_>>::new();
-
-            let mut dealers = Vec::with_capacity(n1 as usize);
-            let mut dealers_encrypted_secret_shares_for_signers: Vec<
-                BTreeMap<u32, EncryptedSecretShare<Secp256k1Sha256>>,
-            > = (0..n1).map(|_| BTreeMap::new()).collect();
-
-            for i in 0..n1 as usize {
-                let (dealer_for_signers, dealer_encrypted_shares_for_signers, _participant_lists) =
-                    Participant::reshare(new_params, &participants_secret_keys[i], &signers, rng)?;
-                dealers.push(dealer_for_signers);
-                dealers_encrypted_secret_shares_for_signers[i] =
-                    dealer_encrypted_shares_for_signers;
-            }
-
-            for i in 0..n2 as usize {
-                let (signer_state, _participant_lists) =
-                    DistributedKeyGeneration::<RoundOne, Secp256k1Sha256>::new(
-                        params,
-                        &signers_dh_secret_keys[i],
-                        signers[i].index,
-                        &dealers,
-                        rng,
-                    )?;
-                signers_states_1.push(signer_state);
-            }
-
-            for (i, shares) in signers_encrypted_secret_shares.iter_mut().enumerate() {
-                let shares_for_signer = dealers_encrypted_secret_shares_for_signers
-                    .iter()
-                    .map(|v| v.get(&(i as u32 + 1)).unwrap().clone())
-                    .collect::<Vec<EncryptedSecretShare<Secp256k1Sha256>>>();
-                *shares = shares_for_signer;
-            }
-
-            for i in 0..n2 as usize {
-                signers_states_2.push(
-                    signers_states_1[i]
-                        .clone()
-                        .to_round_two(&signers_encrypted_secret_shares[i], rng)?
-                        .0,
-                );
-            }
-
-            let mut signers_secret_keys = Vec::<IndividualSigningKey<Secp256k1Sha256>>::new();
-
-            for signers_state in &signers_states_2 {
-                let (_, pi_sk) = signers_state.clone().finish()?;
-
-                signers_secret_keys.push(pi_sk);
-            }
-
-            return Ok((
-                params,
-                participants_secret_keys,
-                group_key,
-                Some(new_params),
-                Some(signers_secret_keys),
-            ));
-        }
-
-        Ok((params, participants_secret_keys, group_key, None, None))
-    }
 
     #[test]
     fn signing_and_verification_single_party() {
